nebexam-backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
drf-spectacular==0.29.0
python-dotenv==1.0.1
Pillow==10.4.0

# Cloudflare R2 storage
django-storages==1.14.6
boto3==1.42.77
